backend/requirements.txt
fastapi
uvicorn[standard]
haystack-ai
sentence-transformers
numpy
faiss-cpu
pandas  # For any data handling
reportlab  # For PDF fallback
python-docx  # For Word exports
huggingface_hub  # For optional commits
sqlalchemy  # For DB (start with SQLite)
python-dotenv
PyJWT  # For JWT authentication
cryptography  # Required by PyJWT for RS256